# Stop caching outputs with `LazyInit` in `CompilerStack`