Depends on #15451.

This is a follow-up to #15451, removing more of the data cached in `CompilerStack`: `metadata`, `abi`, `storageLayout`, `transientStorageLayout`, `userDocumentation`, `devDocumentation`, `generatedSources` and `runtimeGeneratedSources` are now always generated on demand.

This is a test PR to benchmark it see if it's even worth it.

If it's going to be merged, we should perhaps check if some of these aren't used a lot internally and better off cached after all.