@cameel 2024-09-26
### Benchmarks
Results from `external.sh`. #15451 vs this PR.

|         File | Pipeline | Time (before) | Time (after) | Memory (after) | Memory (after) |
|--------------|----------|--------------:|-------------:|---------------:|---------------:|
| openzeppelin |   legacy |         10  s |        9   s |       517  MiB |       448  MiB |
| openzeppelin |       ir |         33  s |        32  s |       497  MiB |       508  MiB |
| uniswap-v4   |   legacy |         21  s |        20  s |       1012 MiB |       964  MiB |
| uniswap-v4   |       ir |         119 s |        119 s |       1494 MiB |       1493 MiB |
| eigenlayer   |   legacy |         74  s |        73  s |       2835 MiB |       2444 MiB |
| eigenlayer   |       ir |         522 s |        522 s |       4457 MiB |       4476 MiB |

Looks like timing is the same. Memory use did fall a little on legacy, but on IR is practically unaffected.