@cameel 2024-09-26
Closing in favor of #15460. Looks like all the gains here are indeed due to generated sources.